Add django-mysql which adds a BUNCH of nice mysql-specific enhancements ot django including performance enhancements, new fields (List, Set, Dict [through DynamicColumn], etc. fields) new mysql-specific cachig options, QuerySet extensions and more.
Removed django-jsonfield (since django-mysql provides that using the JSON data type), requiring a modification of migration 10 in TBAW to remove reference to jsonfield.

Primitives updated to use django-mysql Model (totally compatible with Django's default Model) and for Team to use a ListCharField.

Update addelo to ignore teams or alliances that are inactive for a given year.

Updated requirements with django-mysql, requests-futures (and it's dependencies)