Implement a configuration structure, currently this is a python-based config using modules and classes to provide sections and subsections. This may change in the future.
TemplateStrings now wraps all string methods, making their *apparent* behavior indistinguishable (if done right). Currently unclear if this holds for standard library functions.

Add more string utilities, namely a function to get the fully qualified name of a Type or the Type of an instance and a function that returns a list of template variable names given a formatting string.